RDKB-33758,RDKB-33681: XG is not listed in XI6 after XB reboot
Reason for change: This change would remove the
default addition of multicast routes and arp entries
during system intialization and later
removing them based on authentication status.
Instead routes and arp get added only if the
connected devices are authenticated.
Test Procedure:This is a patch to check
if the xupnp connection issues encountered are due
to the addition and deletion of routes.
Risks: None
Signed-off-by: pmathe686 <philip_mathew@comcast.com>

RDKB-33758: XI6 not listed in MOCA dev after XB reboot

Reason for change: This change reduce the sleep
time of MRD process. This issue is occuring
only randomly and xupnp is supposed to rediscover
even if there is a two min delay in discovering
a device. This fix may help to resolve the issue
though it is masking the xupnp issue of not trying to
rediscover a device after the initial discovery
failure.
Test Procedure:In DIAL network with
home network isolation on restart XB
multiple times and check if the XI (WiFi xupnp client)
device is discovered by MOCA device.
Risks: None
Signed-off-by: pmathe686 <philip_mathew@comcast.com>

Change-Id: I8017211a2c8b149d90e5b6dff594df828a1853ff